backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rafael Ferreira Machado <rafaelf.machado2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
django = "^5.2.6"
djangorestframework = "^3.16.1"
python-decouple = "^3.8"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
drf-spectacular = "^0.28.0"
django-cors-headers = "^4.7.0"
django-health-check = "^3.20.0"
requests = "^2.32.5"
openai = "^1.107.0"
langchain = "^0.3.27"
langchain-core = "^0.3.76"
langchain-community = "^0.3.29"
langchain-openai = "^0.3.33"
faiss-cpu = "^1.12.0"
djangorestframework-simplejwt = "^5.5.1"
langchain-chroma = "^0.2.6"
redis = "^6.4.0"
uvicorn = "^0.35.0"
prometheus-client = "^0.22.1"
django-prometheus = "^2.4.1"
beautifulsoup4 = "^4.13.5"
pypdf = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pytest-cov = "^6.3.0"
pytest-asyncio = "^1.2.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
